Add files via upload
The above project involves creating an animation of a car in Python using a custom graphics module. The animation involves clicking different parts of the road in order to specify the dimensions of the car and also has settings to change the time of the day in the background as well as speed up or slow down the movement of the car.