Add files via upload
This is a game designed in python using a custom graphics module. Various cars are spawned and start travelling across the screen and you need to click on them before they reach the other side of the screen. The game keeps track of what color cars you have clicked on as well as displaying the color of the latest car clicked.